Pembuatan Repack

+++ app/Http/Controllers/Api/Auth/LoginController.php
@@ -35,10 +35,15 @@ class LoginController extends Controller
 
     public function logout(Request $request)
     {
-        $request->user()->currentAccessToken()->delete();
+        if ($request->user()) {
+            $request->user()->currentAccessToken()->delete();
+            return response()->json([
+                'message' => 'Successfully logged out'
+            ]);
+        }
 
         return response()->json([
-            'message' => 'Successfully logged out'
-        ]);
+            'message' => 'Unauthenticated.'
+        ], 401);
     }
 }
